Fixes #13757 - spilt layout helper into two helpers
Layout helper is too large, all the forms logic moved to form helper
